ecommerce/requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.60
botocore==1.17.60
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.14
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
docutils==0.15.2
gensim==3.8.3
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
nltk==3.5
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
sklearn==0.0
smart-open==2.1.1
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
zipp==3.1.0
